# Re-enable shortcut in amd64 decoder